docs: fix examples suspense
