+++ docs/app/global.css
@@ -72,3 +72,10 @@
   font-family: system-ui, -apple-system, BlinkMacSystemFont, "Segoe UI", Roboto,
     Oxygen, Ubuntu, Cantarell, "Open Sans", "Helvetica Neue", sans-serif;
 }
+
+.example-enter {
+  animation: seed-enter;
+  animation-duration: var(--seed-duration-d2);
+  animation-timing-function: var(--seed-timing-function-easing);
+  --seed-enter-opacity: 0;
+}
