+++ docs/components/component-example.tsx
@@ -4,6 +4,7 @@ import * as Index from "./example/index.json";
 
 import ErrorBoundary from "./error-boundary";
 import { CodeBlock } from "./code-block";
+import { ComponentPreview } from "./component-preview";
 
 interface ComponentExampleProps {
   name: string;
@@ -14,16 +15,6 @@ interface ComponentExampleProps {
 export function ComponentExample(props: ComponentExampleProps) {
   const { name } = props;
 
-  const Preview = React.useMemo(() => {
-    const Component = React.lazy(() => import(`./example/${name}.tsx`));
-
-    if (!Component) {
-      return <div>컴포넌트가 존재하지 않습니다.</div>;
-    }
-
-    return <Component />;
-  }, [name]);
-
   const Code = React.useMemo(() => {
     return (Index as Record<string, string>)[name];
   }, [name]);
@@ -31,23 +22,7 @@ export function ComponentExample(props: ComponentExampleProps) {
   if (props.previewOnly) {
     return (
       <React.Suspense fallback={null}>
-        <div
-          className="not-prose example-reset"
-          style={{
-            minHeight: "300px",
-            width: "100%",
-            display: "flex",
-            flexDirection: "column",
-            backgroundColor: "var(--seed-color-bg-layer-default)",
-            justifyContent: "center",
-            alignItems: "center",
-            padding: "10px",
-            borderRadius: ".75rem",
-            border: "1px solid var(--seed-color-stroke-neutral)",
-          }}
-        >
-          {Preview}
-        </div>
+        <ComponentPreview name={name} />
       </React.Suspense>
     );
   }
@@ -56,24 +31,7 @@ export function ComponentExample(props: ComponentExampleProps) {
     <ErrorBoundary>
       <Tabs items={["미리보기", "코드"]}>
         <Tab value="미리보기">
-          <React.Suspense fallback={null}>
-            <div
-              className="not-prose example-reset"
-              style={{
-                minHeight: "300px",
-                width: "100%",
-                display: "flex",
-                flexDirection: "column",
-                backgroundColor: "var(--seed-color-bg-layer-default)",
-                justifyContent: "center",
-                alignItems: "center",
-                padding: "10px",
-                borderRadius: ".75rem",
-              }}
-            >
-              {Preview}
-            </div>
-          </React.Suspense>
+          <ComponentPreview name={name} />
         </Tab>
         <Tab value="코드">
           <CodeBlock lang="tsx" wrapper={{ allowCopy: true }} code={Code} />
