+++ docs/components/component-preview.tsx
@@ -1,0 +1,42 @@
+"use client";
+
+import { ProgressCircle } from "@/registry/ui/progress-circle";
+import * as React from "react";
+import { SimpleReveal } from "simple-reveal";
+
+interface ComponentPreviewProps {
+  name: string;
+}
+
+export function ComponentPreview(props: ComponentPreviewProps) {
+  const { name } = props;
+
+  const Preview = React.useMemo(() => {
+    const Component = React.lazy(() => import(`./example/${name}.tsx`));
+
+    if (!Component) {
+      return <div>컴포넌트가 존재하지 않습니다.</div>;
+    }
+
+    return <Component />;
+  }, [name]);
+
+  return (
+    <div
+      className="not-prose example-reset"
+      style={{
+        minHeight: "300px",
+        width: "100%",
+        display: "flex",
+        flexDirection: "column",
+        backgroundColor: "var(--seed-color-bg-layer-default)",
+        justifyContent: "center",
+        alignItems: "center",
+      }}
+    >
+      <React.Suspense>
+        <div className="example-enter">{Preview}</div>
+      </React.Suspense>
+    </div>
+  );
+}
